refactor: 그래프 뷰 백그라운드 이벤트 삭제, 사이드바 구조 수정
